Moved jpg moving logic to update script
